requirements.txt
requests>=2.25.1
pytest>=6.2.5
python-dotenv>=0.19.0 